Fix 3 blockers da campanha metric-learning (sem mudança numérica)
Correções encontradas rodando a campanha de 270 experimentos:

- teacher_models.load_teacher: strict=False do PyTorch não ignora
  size mismatch de uma chave presente nos dois lados. A cabeça
  classificadora (m.fc / classifier[2]) muda de nº de classes entre
  datasets e não é usada em metric learning -> filtra params com shape
  incompatível antes do load_state_dict (carrega só backbone/embedding).

- graph_rkd.embeddings.mds_spectral_embedding: torch.linalg.eigvalsh não
  tem kernel CUDA para Half; sob AMP quebrava com
  "linalg_eigh_cuda not implemented for 'Half'". Calcula em float32 e
  converte de volta (a decomposição espectral precisa de precisão plena).

- distill_metric / finetune_metric / train_metric_baseline: default de
  --workers 8 -> 4. Com 2 workers paralelos x 8 DataLoader workers x 4
  loaders persistentes havia oversubscription -> deadlock de futex no
  DataLoader. 4 elimina os hangs sem alterar nada numérico.

+++ RKD/graph_rkd/embeddings.py
@@ -119,7 +119,9 @@ def mds_spectral_embedding(D, normalize=False, jitter=0.0):
     B = 0.5 * (B + B.transpose(-2, -1))             # força simetria (estabilidade)
     if jitter > 0:
         B = B + jitter * torch.eye(N, device=D.device, dtype=D.dtype)
-    eig = torch.linalg.eigvalsh(B)                  # crescente
+    # eigvalsh não tem kernel CUDA p/ Half (AMP); calcula em float32 e converte
+    # de volta. A decomposição espectral precisa de precisão plena de qualquer forma.
+    eig = torch.linalg.eigvalsh(B.float()).to(B.dtype)   # crescente
     eig = eig.flip(-1)                              # decrescente
     return eig.squeeze(0) if single else eig
